feat(gradle): merge gradle-lite into gradle (#10132)
Removes standalone gradle-lite manager and merges it into the gradle manager as the default behavior, while adding deepExtract as an option to re-enable previous behavior.

Closes #10091

BREAKING CHANGE: Gradle extraction now defaults to JS-based parsing (previously "gradle-lite")
