+++ lib/config/migration.ts
@@ -580,6 +580,23 @@ export function migrateConfig(
         }
       }
     }
+    if (is.nonEmptyArray(migratedConfig.matchManagers)) {
+      if (migratedConfig.matchManagers.includes('gradle-lite')) {
+        if (!migratedConfig.matchManagers.includes('gradle')) {
+          migratedConfig.matchManagers.push('gradle');
+        }
+        migratedConfig.matchManagers = migratedConfig.matchManagers.filter(
+          (manager) => manager !== 'gradle-lite'
+        );
+      }
+    }
+    if (is.nonEmptyObject(migratedConfig['gradle-lite'])) {
+      migratedConfig.gradle = mergeChildConfig(
+        migratedConfig.gradle || {},
+        migratedConfig['gradle-lite']
+      );
+    }
+    delete migratedConfig['gradle-lite'];
     const isMigrated = !dequal(config, migratedConfig);
     if (isMigrated) {
       // recursive call in case any migrated configs need further migrating
