+++ lib/config/presets/common.ts
@@ -3,10 +3,12 @@ export const removedPresets = {
   ':automergeBranchPush': ':automergeBranch',
   ':base': 'config:base',
   ':app': 'config:js-app',
+  ':enableGradleLite': null,
   ':js-app': 'config:js-app',
   ':library': 'config:js-lib',
   ':masterIssue': ':dependencyDashboard',
   ':masterIssueApproval': ':dependencyDashboardApproval',
+  ':switchToGradleLite': null,
   ':unpublishSafe': 'npm:unpublishSafe',
   'config:application': 'config:js-app',
   'config:base-js': 'config:base',
