+++ lib/manager/api.ts
@@ -24,7 +24,6 @@ import * as gitlabci from './gitlabci';
 import * as gitlabciInclude from './gitlabci-include';
 import * as gomod from './gomod';
 import * as gradle from './gradle';
-import * as gradleLite from './gradle-lite';
 import * as gradleWrapper from './gradle-wrapper';
 import * as helmRequirements from './helm-requirements';
 import * as helmValues from './helm-values';
@@ -92,7 +91,6 @@ api.set('gitlabci', gitlabci);
 api.set('gitlabci-include', gitlabciInclude);
 api.set('gomod', gomod);
 api.set('gradle', gradle);
-api.set('gradle-lite', gradleLite);
 api.set('gradle-wrapper', gradleWrapper);
 api.set('helm-requirements', helmRequirements);
 api.set('helm-values', helmValues);
