--- lib/manager/gradle-lite/index.ts
@@ -1,10 +1,0 @@
-import * as gradleVersioning from '../../versioning/gradle';
-
-export { extractAllPackageFiles } from './extract';
-export { updateDependency } from './update';
-
-export const defaultConfig = {
-  fileMatch: ['(^|/)gradle.properties$', '\\.gradle(\\.kts)?$'],
-  versioning: gradleVersioning.id,
-  enabled: false,
-};
