+++ lib/manager/gradle/deep/__fixtures__/MultiProjectUpdatesReport.json
