+++ lib/manager/gradle/deep/__fixtures__/build.gradle.example1
