+++ lib/manager/gradle/deep/__fixtures__/gradle-wrappers/5/gradle/wrapper/gradle-wrapper.properties
