+++ lib/manager/gradle/deep/__fixtures__/gradle-wrappers/6/gradlew
