+++ lib/manager/gradle/deep/__fixtures__/minimal-project/settings.gradle
