+++ lib/manager/gradle/deep/__fixtures__/updatesReport.json
