+++ lib/manager/gradle/deep/__fixtures__/updatesReportEmpty.json
