+++ lib/manager/gradle/deep/__testutil__/gradle.ts
