+++ lib/manager/gradle/deep/build-gradle.spec.ts
@@ -1,4 +1,4 @@
-import { getName } from '../../../test/util';
+import { getName } from '../../../../test/util';
 import {
   collectVersionVariables,
   init,
