+++ lib/manager/gradle/readme.md
@@ -1,1 +1,5 @@
-Renovate extracts Gradle dependencies by calling a custom Gradle script. The Gradle binary is then used to extract Maven-type dependencies.
+The `gradle` manager's default behavior uses a custom parser written in JavaScript, similar to many others managers.
+It was initially known as `gradle-lite` but is now integrated into the `gradle` manager and used as default.
+
+If `deepExtract` is configured to `true`, Renovate instead extracts Gradle dependencies by calling a custom Gradle script.
+The `gradle` binary is then used to extract Maven-type dependencies.
