+++ lib/manager/gradle/shallow/common.ts
@@ -1,4 +1,4 @@
-export { MAVEN_REPO } from '../../datasource/maven/common';
+export { MAVEN_REPO } from '../../../datasource/maven/common';
 
 export const JCENTER_REPO = 'https://jcenter.bintray.com/';
 export const GOOGLE_REPO = 'https://dl.google.com/android/maven2/';
