+++ lib/manager/gradle/shallow/extract.spec.ts
@@ -1,7 +1,7 @@
-import { fs, getName } from '../../../test/util';
-import { extractAllPackageFiles } from '.';
+import { extractAllPackageFiles } from '..';
+import { fs, getName } from '../../../../test/util';
 
-jest.mock('../../util/fs');
+jest.mock('../../../util/fs');
 
 function mockFs(files: Record<string, string>): void {
   fs.readLocalFile.mockImplementation((fileName: string): Promise<string> => {
