+++ lib/manager/gradle/shallow/index.ts
@@ -1,0 +1,2 @@
+export { extractAllPackageFiles } from './extract';
+export { updateDependency } from './update';
