+++ lib/manager/gradle/shallow/tokenizer.spec.ts
@@ -1,4 +1,4 @@
-import { getName } from '../../../test/util';
+import { getName } from '../../../../test/util';
 import { TokenType } from './common';
 import { extractRawTokens, tokenize } from './tokenizer';
